Fetch video page data from Sanity
