+++ src/pages/video/[productionSlug]/player.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables.scss';
+@import '../../../variables.scss';
 
 .wrapper {
   position: absolute;
